User management
The main fix in this commit is to allow users to log in correctly and for users to be registered correctly (There was an issue with the reg date). Also, when the user logs in, there was no redirection to a users/me page. Another improvement was expanding the README to allow running using docker, local, live reload and visual studio code debugging. Environment file has been created